split swift reader and writer
